chore: move Exceptions to root directory.
